connect App to firebase auth; display splash screen while loading auth state
